Samarth: made apis

+++ src/controller/comment.controller.js
@@ -1,9 +1,11 @@
 import { StatusCodes } from "http-status-codes";
-
-import Comment from "../models/Comment.js";
 import mongoose from "mongoose";
+
+import Activity from "../models/Activity.js";
+import Comment from "../models/Comment.js";
 import Card from "../models/Card.js";
 import ApiError from "../utils/ApiError.js";
+import { populate } from "dotenv";
 
 export const getAllComments = async (req, res, next) => {
   const { cardId } = req.params;
@@ -16,6 +18,9 @@ export const getAllComments = async (req, res, next) => {
       path: "comments",
       populate: {
         path: "author",
+        populate: {
+          path: "avatar",
+        },
       },
     });
 
@@ -33,10 +38,12 @@ export const getAllComments = async (req, res, next) => {
 
 export const createComment = async (req, res, next) => {
   const { text, cardId } = req.body;
-  const userId = req.user.userId;
+  const user = req.user;
 
   if (!text.trim()) {
-    return next(new ApiError("comment text is required!", StatusCodes.BAD_REQUEST));
+    return next(
+      new ApiError("comment text is required!", StatusCodes.BAD_REQUEST)
+    );
   }
 
   if (!mongoose.Types.ObjectId.isValid(cardId)) {
@@ -52,12 +59,22 @@ export const createComment = async (req, res, next) => {
 
     const newComment = await Comment.create({
       text,
-      author: userId,
+      author: user.userId,
       card: cardId,
+      board: card.board,
     });
 
     card.comments.push(newComment._id);
     card.save();
+
+    await Activity.create({
+      type: "COMMENT_ADDED",
+      board: card.board,
+      column: card.column,
+      card: cardId,
+      user: user.userId,
+      description: `${user.name} added comment on "${card.title}"`,
+    });
 
     return res.status(StatusCodes.OK).json({ message: "comment created!" });
   } catch (error) {
@@ -67,6 +84,8 @@ export const createComment = async (req, res, next) => {
 
 export const deleteComment = async (req, res, next) => {
   const { commentId } = req.params;
+  const user = req.user;
+
   if (!mongoose.Types.ObjectId.isValid(commentId)) {
     return next(new ApiError("Invalid comment ID.", StatusCodes.BAD_REQUEST));
   }
@@ -76,6 +95,17 @@ export const deleteComment = async (req, res, next) => {
     if (!comment) {
       return next(new ApiError("comment not found!", StatusCodes.NOT_FOUND));
     }
+
+    if (comment.author.toString() !== user.userId.toString()) {
+      return next(
+        new ApiError(
+          "You can not delete this comment.",
+          StatusCodes.UNAUTHORIZED
+        )
+      );
+    }
+
+    await Activity.deleteMany({ comment: commentId });
 
     await comment.deleteOne();
     return res.status(StatusCodes.OK).json({ message: `comment deleted!` });
